resume-parser/requirements.txt
blinker==1.4
blis==0.2.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cymem==2.0.2
en-core-web-sm==2.1.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
MarkupSafe==1.1.1
murmurhash==1.0.2
numpy==1.16.2
pandas==0.24.2
pdfminer.six==20181108
plac==0.9.6
preshed==2.0.1
pycryptodome==3.8.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
six==1.12.0
sortedcontainers==2.1.0
spacy==2.1.3
SQLAlchemy==1.3.2
srsly==0.0.5
thinc==7.0.4
tqdm==4.31.1
urllib3==1.24.1
wasabi==0.2.1
Werkzeug==0.15.2
Whoosh==2.7.4
WTForms==2.2.1
